% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol-facets.R
\name{io_statistics_facets}
\alias{io_statistics_facets}
\alias{InputStatisticsInputDatasetFacet}
\alias{OutputStatisticsOutputDatasetFacet}
\title{Input and output statistics facets}
\usage{
InputStatisticsInputDatasetFacet(
  row_count = NULL,
  size = NULL,
  file_count = NULL,
  producer = OPENLINEAGE_PRODUCER
)

OutputStatisticsOutputDatasetFacet(
  row_count = NULL,
  size = NULL,
  file_count = NULL,
  producer = OPENLINEAGE_PRODUCER
)
}
\arguments{
\item{row_count}{The optional number of rows read or written.}

\item{size}{The optional number of bytes read or written.}

\item{file_count}{The optional number of files read or written.}

\item{producer}{A URI identifying the facet producer.}
}
\value{
A typed input- or output-dataset facet S7 object.
}
\description{
Input and output statistics facets
}
