% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{openlineage_constants}
\alias{openlineage_constants}
\alias{OPENLINEAGE_SCHEMA_VERSION}
\alias{OPENLINEAGE_SCHEMA_URL}
\alias{OPENLINEAGE_RUN_EVENT_SCHEMA_URL}
\alias{OPENLINEAGE_PRODUCER}
\title{OpenLineage protocol constants}
\format{
Character scalars.
}
\usage{
OPENLINEAGE_SCHEMA_VERSION

OPENLINEAGE_SCHEMA_URL

OPENLINEAGE_RUN_EVENT_SCHEMA_URL

OPENLINEAGE_PRODUCER
}
\description{
Constants used by \code{openlineage} when constructing OpenLineage events.
\code{OPENLINEAGE_SCHEMA_URL} identifies the schema document, while
\code{OPENLINEAGE_RUN_EVENT_SCHEMA_URL} identifies its \code{RunEvent} definition.
}
\examples{
OPENLINEAGE_SCHEMA_VERSION
OPENLINEAGE_SCHEMA_URL
OPENLINEAGE_RUN_EVENT_SCHEMA_URL
OPENLINEAGE_PRODUCER
}
