% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol-facets.R
\name{tags_facets}
\alias{tags_facets}
\alias{TagsJobFacet}
\alias{TagsRunFacet}
\alias{TagsDatasetFacet}
\title{Typed tags facets}
\usage{
TagsJobFacet(tags = list(), producer = OPENLINEAGE_PRODUCER, deleted = NULL)

TagsRunFacet(tags = list(), producer = OPENLINEAGE_PRODUCER)

TagsDatasetFacet(
  tags = list(),
  producer = OPENLINEAGE_PRODUCER,
  deleted = NULL
)
}
\arguments{
\item{tags}{An unnamed list created with \code{ol_tag()}.}

\item{producer}{A URI identifying the facet producer.}

\item{deleted}{Whether a job or dataset facet is being deleted.}
}
\value{
A typed tags facet S7 object.
}
\description{
Typed tags facets
}
