% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openlineage-package.R
\docType{package}
\name{openlineage-package}
\alias{openlineage}
\alias{openlineage-package}
\title{openlineage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/openlineage/}
  \item \url{https://github.com/pedrobtz/openlineage}
  \item Report bugs at \url{https://github.com/pedrobtz/openlineage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
